Fix Harga di Tabel Barang

+++ database/migrations/2025_10_20_030932_create_barang_table.php
@@ -23,7 +23,7 @@ class CreateBarangTable extends Migration
             $table->unsignedBigInteger('id_tipe');
             $table->unsignedBigInteger('id_status');
             $table->date('tanggal_masuk');
-            $table->decimal('harga', 15, 2)->nullable();
+            $table->bigInteger('harga')->nullable();
             $table->string('lokasi', 100)->nullable();
             $table->text('keterangan')->nullable();
             $table->timestamps();
